// src/model/out/character/elementalBurst.ts
import { UnlockTypes } from "@prisma/client";
import { DescriptionOut } from "./description";
import { UpgradeItemOut } from "./upgradeItem";

export interface ElementalBurstOut {
    unlock: UnlockTypes | null;
    name: string;
    note: string;
    descriptions: DescriptionOut[];
    upgrades: UpgradeItemOut[];
}